// src/app/actions/types.ts
export enum ActionType {
    SELECT_TOOL = 'SELECT_TOOL',
    MOVE_ITINERARY_POINT = 'MOVE_ITINERARY_POINT',
    ADD_ITINERARY_POINT = 'ADD_ITINERARY_POINT',
    UPDATE_ITINERARY_POINT = 'UPDATE_ITINERARY_POINT',
    REMOVE_ITINERARY_POINT = 'REMOVE_ITINERARY_POINT',
    GO_TO = 'GO_TO',
    GO_FROM = 'GO_FROM',
    FETCH_ROUTES = 'FETCH_ROUTES',
    FETCH_ROUTES_SUCCESS = 'FETCH_ROUTES_SUCCESS',
    FETCH_ROUTES_FAILURE = 'FETCH_ROUTES_FAILURE',
    RESET_ROUTES = 'RESET_ROUTES',
    FETCH_POINTS_OF_INTEREST = 'FETCH_POINTS_OF_INTEREST',
    FETCH_POINTS_OF_INTEREST_SUCCESS = 'FETCH_POINTS_OF_INTEREST_SUCCESS',
    FETCH_POINTS_OF_INTEREST_FAILURE = 'FETCH_POINTS_OF_INTEREST_FAILURE',
    FETCH_SEARCH_RESULT = 'FETCH_SEARCH_RESULT',
    FETCH_SEARCH_RESULT_SUCCESS = 'FETCH_SEARCH_RESULT_SUCCESS',
    FETCH_SEARCH_RESULT_FAILURE = 'FETCH_SEARCH_RESULT_FAILURE',
    SELECT_SEARCH_RESULT = 'SELECT_SEARCH_RESULT',
    CHANGE_THEME = 'CHANGE_THEME',
    ZOOM_IN = 'ZOOM_IN',
    ZOOM_OUT = 'ZOOM_OUT'
}